Add read_data method to ASReviewProject (#1621)
