+++ asreview/project.py
@@ -28,6 +28,7 @@ __all__ = [
 import json
 import logging
 import os
+import pickle
 import shutil
 import tempfile
 import time
@@ -40,6 +41,7 @@ from uuid import uuid4
 
 import jsonschema
 import numpy as np
+import pandas as pd
 from filelock import FileLock
 from scipy.sparse import csr_matrix
 from scipy.sparse import load_npz
@@ -51,6 +53,7 @@ from asreview.config import PROJECT_MODE_SIMULATE
 from asreview.config import PROJECT_MODES
 from asreview.config import SCHEMA
 from asreview.data import ASReviewData
+from asreview.exceptions import CacheDataError
 from asreview.state.errors import StateNotFoundError
 from asreview.state.sqlstate import SQLiteState
 from asreview.utils import asreview_path
@@ -365,6 +368,69 @@ class ASReviewProject:
             Path(self.project_path, "reviews").iterdir()
         ):
             self.delete_review()
+
+    def _read_data_from_cache(self, version_check=True):
+
+        fp_data = Path(self.project_path, "data", self.config["dataset_path"])
+        fp_data_pickle = Path(fp_data).with_suffix(fp_data.suffix + ".pickle")
+
+        try:
+            with open(fp_data_pickle, "rb") as f_pickle_read:
+                data_obj, data_obj_version = pickle.load(f_pickle_read)
+
+            if not isinstance(data_obj.df, pd.DataFrame):
+                raise ValueError()
+
+            if (not version_check) or (get_versions()["version"] == data_obj_version):
+                return data_obj
+
+        except FileNotFoundError:
+            pass
+        except Exception as err:
+            logging.error(f"Error reading cache file: {err}")
+            try:
+                os.remove(fp_data_pickle)
+            except FileNotFoundError:
+                pass
+
+        raise CacheDataError()
+
+    def read_data(self, use_cache=True, save_cache=True):
+        """Get ASReviewData object from file.
+
+        Parameters
+        ----------
+        use_cache: bool
+            Use the pickle file if available.
+        save_cache: bool
+            Save the file to a pickle file if not available.
+
+        Returns
+        -------
+        ASReviewData:
+            The data object for internal use in ASReview.
+
+        """
+
+        try:
+            fp_data = Path(self.project_path, "data", self.config["dataset_path"])
+        except Exception:
+            raise FileNotFoundError("Dataset not found")
+
+        if use_cache:
+            try:
+                return self._read_data_from_cache(fp_data)
+            except CacheDataError:
+                pass
+
+        data_obj = ASReviewData.from_file(fp_data)
+
+        if save_cache:
+            fp_data_pickle = Path(fp_data).with_suffix(fp_data.suffix + ".pickle")
+            with open(fp_data_pickle, "wb") as f_pickle:
+                pickle.dump((data_obj, get_versions()["version"]), f_pickle)
+
+        return data_obj
 
     def clean_tmp_files(self):
         """Clean temporary files in a project.
